dom/nodes/Element-childElementCount-nochild-svg.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="100%" height="100%" version="1.1" viewBox="0 0 400 400"><title>childElementCount</title><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><text x="200" y="40" fill="#000" font-size="25" text-anchor="middle">Test of childElementCount with No Child Element Nodes</text><text id="parentEl" x="200" y="70" fill="#000" font-size="20" font-weight="bold" text-anchor="middle">Test</text><h:script><![CDATA[
test(function() {
  var parentEl = document.getElementById("parentEl")
  assert_equals(parentEl.childElementCount, 0)
})
]]></h:script></svg>
